producer/requirements.txt
mysql-connector-python==8.1.0
confluent-kafka==2.3.0
python-dotenv==1.0.1
